libs/utils/analysis: add TASKS module
This patch move the tasks's signals analysis code, originally part of the
trace_analysis.py module, to be a new TraceAnalysis module with dynamic
registration via the AnalysisRegister.
